feat(App): Obtain GIFs via HTTP request using Axios
