+++ src/gifs/intearfaces/gif.interface.ts
@@ -1,0 +1,8 @@
+
+export interface Gif {
+  id: string;
+  title: string;
+  url: string;
+  width: number;
+  height: number;
+}
